COM-1677 moved to mixin
