+++ src/modules/client/pages/ni/courses/LearnersDirectory.vue
@@ -35,15 +35,14 @@ import { frPhoneNumber } from '@helpers/vuelidateCustomVal';
 import { required, requiredIf, email } from 'vuelidate/lib/validators';
 import { DEFAULT_AVATAR } from '@data/constants';
 import {
-  formatIdentity,
   formatPhoneForPayload,
   removeDiacritics,
   removeEmptyProps,
-  sortStrings,
   clear,
 } from '@helpers/utils';
 import { NotifyPositive, NotifyNegative, NotifyWarning } from '@components/popup/notify';
 import { userMixin } from '@mixins/userMixin';
+import { learnerDirectoryMixin } from '@mixins/learnerDirectoryMixin';
 
 export default {
   metaInfo: { title: 'Répertoire apprenants' },
@@ -53,43 +52,12 @@ export default {
     'ni-table-list': TableList,
     'learner-creation-modal': LearnerCreationModal,
   },
-  mixins: [userMixin],
+  mixins: [userMixin, learnerDirectoryMixin],
   data () {
     return {
-      tableLoading: false,
       loading: false,
       learnerList: [],
       searchStr: '',
-      pagination: { sortBy: 'name', descending: false, page: 1, rowsPerPage: 15 },
-      columns: [
-        {
-          name: 'name',
-          label: 'Nom',
-          field: row => row.learner,
-          align: 'left',
-          sortable: true,
-          sort: (a, b) => sortStrings(a.lastname, b.lastname),
-          style: 'min-width: 200px; width: 35%',
-        },
-        {
-          name: 'blendedCoursesCount',
-          label: 'Formations mixtes',
-          field: 'blendedCoursesCount',
-          align: 'center',
-          sortable: true,
-          style: 'min-width: 110px; width: 15%',
-          sort: (a, b) => b - a,
-        },
-        {
-          name: 'eLearningCoursesCount',
-          label: 'Formations eLearning',
-          field: 'eLearningCoursesCount',
-          align: 'center',
-          sortable: true,
-          style: 'min-width: 110px; width: 15%',
-          sort: (a, b) => b - a,
-        },
-      ],
       learnerCreationModal: false,
       learnerCreationModalLoading: false,
       identityStep: false,
@@ -121,22 +89,6 @@ export default {
     },
     updateSearch (value) {
       this.searchStr = value;
-    },
-    formatRow (user) {
-      const formattedName = formatIdentity(user.identity, 'FL');
-
-      return {
-        learner: {
-          _id: user._id,
-          fullName: formattedName,
-          lastname: user.identity.lastname,
-          picture: user.picture ? user.picture.link : null,
-          noDiacriticsName: removeDiacritics(formattedName),
-        },
-        company: user.company ? user.company.name : 'N/A',
-        blendedCoursesCount: user.blendedCoursesCount,
-        eLearningCoursesCount: user.eLearningCoursesCount,
-      };
     },
     async getLearnerList () {
       try {
